+++ src/main/webapp/jqmng/activate.js
@@ -1,4 +1,4 @@
-define(['jquery', 'jqmng/globalScope'], function($, globalScope) {
+define(['jquery', 'angular'], function($, angular) {
     /*
      * Service for page navigation.
      * Parameters (see $.mobile.changePage)
@@ -16,6 +16,34 @@ define(['jquery', 'jqmng/globalScope'], function($, globalScope) {
             $.mobile.changePage.call($.mobile, pageId, transition);
         }
     }
+
+    /**
+     * Helper function to put the navigation part out of the controller into the page.
+     * @param self
+     * @param result
+     * @param trueCase
+     * @param falseCase
+     */
+    angular.Object.activate = function(self, result, trueCase, falseCase) {
+        if (result && result.then) {
+            result.then(function() {
+                activate(trueCase);
+            }, function() {
+                if (falseCase) {
+                    activate(falseCase);
+                }
+            });
+        } else {
+            if (result!==false) {
+                activate(trueCase);
+            } else {
+                if (falseCase) {
+                    activate(falseCase);
+                }
+            }
+        }
+    };
+
 
     return {
         activate: activate